Fixed compilation errors from switch in IGeometry to using size_t

+++ SIMPLVtkLib/SIMPLBridge/VSEdgeGeom.cpp
@@ -122,7 +122,7 @@ void VSEdgeGeom::GetCellPoints(vtkIdType cellId, vtkIdList* ptIds)
 {
   const int numVerts = 2;
 
-  int64_t verts[numVerts];
+  size_t verts[numVerts];
   m_Geom->getVertsAtEdge(cellId, verts);
 
   ptIds->SetNumberOfIds(numVerts);
@@ -139,7 +139,7 @@ void VSEdgeGeom::GetPointCells(vtkIdType ptId, vtkIdList* cellIds)
 {
   ElementDynamicList::Pointer elementsContainingList = m_Geom->getElementsContainingVert();
 
-  DynamicListArray<uint16_t, int64_t>::ElementList listArray = elementsContainingList->getElementList(ptId);
+  DynamicListArray<uint16_t, size_t>::ElementList listArray = elementsContainingList->getElementList(ptId);
 
   cellIds->SetNumberOfIds(listArray.ncells);
   for(int i = 0; i < listArray.ncells; i++)
